solved Sumit and Home Issue

+++ src/components/post-form/PostForm.jsx
@@ -1,9 +1,304 @@
-import React, { useCallback } from "react";
+// import React, { useCallback } from "react";
+// import { useForm } from "react-hook-form";
+// import { Button, Input, RTE, Select } from "..";
+// import appwriteService from "../../appwrite/config";
+// import { useNavigate } from "react-router-dom";
+// import { useSelector } from "react-redux";
+
+// export default function PostForm({ post }) {
+//     const { register, handleSubmit, watch, setValue, control, getValues } = useForm({
+//         defaultValues: {
+//             title: post?.title || "",
+//             slug: post?.$id || "",
+//             content: post?.content || "",
+//             status: post?.status || "active",
+//         },
+//     });
+
+//     const navigate = useNavigate();
+//     const userData = useSelector((state) => state.auth.userData);
+
+//     const submit = async (data) => {
+//         if (post) {
+//             const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+
+//             if (file) {
+//                 appwriteService.deleteFile(post.featuredImage);
+//             }
+
+//             const dbPost = await appwriteService.updatePost(post.$id, {
+//                 ...data,
+//                 featuredImage: file ? file.$id : undefined,
+//             });
+
+//             if (dbPost) {
+//                 navigate(`/post/${dbPost.$id}`);
+//             }
+//         } else {
+//             const file = await appwriteService.uploadFile(data.image[0]);
+
+//             if (file) {
+//                 const fileId = file.$id;
+//                 data.featuredImage = fileId;
+//                 const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
+
+//                 if (dbPost) {
+//                     navigate(`/post/${dbPost.$id}`);
+//                 }
+//             }
+//         }
+//     };
+
+//     const slugTransform = useCallback((value) => {
+//         if (value && typeof value === "string")
+//             return value
+//                 .trim()
+//                 .toLowerCase()
+//                 .replace(/[^a-zA-Z\d\s]+/g, "-")
+//                 .replace(/\s/g, "-");
+
+//         return "";
+//     }, []);
+
+//     React.useEffect(() => {
+//         const subscription = watch((value, { name }) => {
+//             if (name === "title") {
+//                 setValue("slug", slugTransform(value.title), { shouldValidate: true });
+//             }
+//         });
+
+//         return () => subscription.unsubscribe();
+//     }, [watch, slugTransform, setValue]);
+
+//     return (
+//         <form onSubmit={handleSubmit(submit)} className="flex flex-wrap">
+//             <div className="w-2/3 px-2">
+//                 <Input
+//                     label="Title :"
+//                     placeholder="Title"
+//                     className="mb-4"
+//                     {...register("title", { required: true })}
+//                 />
+//                 <Input
+//                     label="Slug :"
+//                     placeholder="Slug"
+//                     className="mb-4"
+//                     {...register("slug", { required: true })}
+//                     onInput={(e) => {
+//                         setValue("slug", slugTransform(e.currentTarget.value), { shouldValidate: true });
+//                     }}
+//                 />
+//                 <RTE label="Content :" name="content" control={control} defaultValue={getValues("content")} />
+//             </div>
+//             <div className="w-1/3 px-2">
+//                 <Input
+//                     label="Featured Image :"
+//                     type="file"
+//                     className="mb-4"
+//                     accept="image/png, image/jpg, image/jpeg, image/gif"
+//                     {...register("image", { required: !post })}
+//                 />
+//                 {post && (
+//                     <div className="w-full mb-4">
+//                         <img
+//                             src={appwriteService.getFilePreview(post.featuredImage)}
+//                             alt={post.title}
+//                             className="rounded-lg"
+//                         />
+//                     </div>
+//                 )}
+//                 <Select
+//                     options={["active", "inactive"]}
+//                     label="Status"
+//                     className="mb-4"
+//                     {...register("status", { required: true })}
+//                 />
+//                 <Button type="submit" bgColor={post ? "bg-green-500" : undefined} className="w-full">
+//                     {post ? "Update" : "Submit"}
+//                 </Button>
+//             </div>
+//         </form>
+//     );
+// }
+
+
+
+
+
+
+// import React, { useCallback, useEffect, useState } from "react";
+// import { useForm } from "react-hook-form";
+// import { Button, Input, RTE, Select } from "..";
+// import appwriteService from "../../appwrite/config";
+// import { useNavigate } from "react-router-dom";
+// import { useSelector } from "react-redux";
+// import { ImSpinner2 } from "react-icons/im";
+
+// export default function PostForm({ post }) {
+//     const { register, handleSubmit, watch, setValue, control, getValues } = useForm({
+//         defaultValues: {
+//             title: post?.title || "",
+//             slug: post?.$id || "",
+//             content: post?.content || "",
+//             status: post?.status || "active",
+//         },
+//     });
+
+//     const navigate = useNavigate();
+//     const [userData, setUserData] = useState(useSelector((state) => state.auth.userData));
+//     // const userData = useSelector((state) => state.auth.userData);
+//     const [loader, setLoader] = useState(false);
+//     const [error, setError] = useState("");
+
+//     useEffect(() => {
+//         const unsubscribe = useSelector((state) => {
+//             setUserData(state.auth.userData);
+//         });
+//         return () => unsubscribe(); // Cleanup subscription
+//     }, [useSelector]);
+
+
+//     const submit = async (data) => {
+//         setError("");
+//         setLoader(true);
+//         try{
+//             console.log("data", data);
+//             console.log("post", post);
+//             if (post) {
+//                 const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+    
+//                 if (file) {
+//                     appwriteService.deleteFile(post.featuredImage);
+//                 }
+    
+//                 const dbPost = await appwriteService.updatePost(post.$id, {
+//                     ...data,
+//                     featuredImage: file ? file.$id : undefined,
+//                 });
+    
+//                 if (dbPost) {
+//                     navigate(`/post/${dbPost.$id}`);
+//                 }
+//             } 
+//             else {
+//                 if (!data.image[0]) {
+//                     console.error("No image provided");
+//                     return;
+//                 }
+//                 const file = await appwriteService.uploadFile(data.image[0]);
+//                 console.log("file", file);
+//                 console.log("userData", userData);
+    
+//                 if (file && userData) {
+//                     const fileId = file.$id;
+//                     data.featuredImage = fileId;
+//                     const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
+    
+//                     if (dbPost) {
+//                         navigate(`/post/${dbPost.$id}`);
+//                     }
+//                 }
+//             }
+//         }   catch (error) {
+//                 console.log("Error on PostForm on submit function", error);
+//                 setError(error.message);
+//         }   finally {
+//                 setLoader(false);
+//         }
+        
+//     };
+
+//     const slugTransform = useCallback((value) => {
+//         if (value && typeof value === "string")
+//             return value
+//                 .trim()
+//                 .toLowerCase()
+//                 .replace(/[^a-zA-Z\d\s]+/g, "-")
+//                 .replace(/\s/g, "-");
+
+//         return "";
+//     }, []);
+
+//     useEffect(() => {
+//         const subscription = watch((value, { name }) => {
+//             if (name === "title") {
+//                 setValue("slug", slugTransform(value.title), { shouldValidate: true });
+//             }
+//         });
+
+//         return () => subscription.unsubscribe();
+//     }, [watch, slugTransform, setValue]);
+
+//     return (
+//         <form onSubmit={handleSubmit(submit)} className="flex flex-wrap">
+//             <div className="w-2/3 px-2">
+//                 <Input
+//                     label="Title :"
+//                     placeholder="Title"
+//                     className="mb-4"
+//                     {...register("title", { required: true })}
+//                 />
+//                 <Input
+//                     label="Slug :"
+//                     placeholder="Slug"
+//                     className="mb-4"
+//                     {...register("slug", { required: true })}
+//                     onInput={(e) => {
+//                         setValue("slug", slugTransform(e.currentTarget.value), { shouldValidate: true });
+//                     }}
+//                 />
+//                 <RTE label="Content :" name="content" control={control} defaultValue={getValues("content")} />
+//             </div>
+//             <div className="w-1/3 px-2">
+//                 <Input
+//                     label="Featured Image :"
+//                     type="file"
+//                     className="mb-4"
+//                     accept="image/png, image/jpg, image/jpeg, image/gif"
+//                     {...register("image", { required: !post })}
+//                 />
+//                 {post && (
+//                     <div className="w-full mb-4">
+//                         <img
+//                             src={appwriteService.getFilePreview(post.featuredImage)}
+//                             alt={post.title}
+//                             className="rounded-lg"
+//                         />
+//                     </div>
+//                 )}
+//                 <Select
+//                     options={["active", "inactive"]}
+//                     label="Status"
+//                     className="mb-4"
+//                     {...register("status", { required: true })}
+//                 />
+//                 <Button type="submit" bgColor={post ? "bg-green-500" : undefined} className="w-full">
+//                     {loader ? (
+//                         <ImSpinner2 className="animate-spin mx-auto" />
+//                     ) : post ? (
+//                       "Update"
+//                     ) : (
+//                       "Submit"
+//                     )}
+//                 </Button>
+//                 {error && <p className="text-red-400">{error}</p>}
+//             </div>
+//         </form>
+//     );
+// }
+
+
+
+
+
+
+import React, { useCallback, useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { Button, Input, RTE, Select } from "..";
 import appwriteService from "../../appwrite/config";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
+import { ImSpinner2 } from "react-icons/im";
 
 export default function PostForm({ post }) {
     const { register, handleSubmit, watch, setValue, control, getValues } = useForm({
@@ -17,35 +312,60 @@ export default function PostForm({ post }) {
 
     const navigate = useNavigate();
     const userData = useSelector((state) => state.auth.userData);
+    const [loader, setLoader] = useState(false);
+    const [error, setError] = useState("");
 
     const submit = async (data) => {
-        if (post) {
-            const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
+        if (!userData) {
+            setError("User data not available. Please try again.");
+            return;
+        }
 
-            if (file) {
-                appwriteService.deleteFile(post.featuredImage);
-            }
+        setError("");
+        setLoader(true);
+        try {
+            console.log("data", data);
+            console.log("post", post);
+            if (post) {
+                const file = data.image[0] ? await appwriteService.uploadFile(data.image[0]) : null;
 
-            const dbPost = await appwriteService.updatePost(post.$id, {
-                ...data,
-                featuredImage: file ? file.$id : undefined,
-            });
+                if (file) {
+                    appwriteService.deleteFile(post.featuredImage);
+                }
 
-            if (dbPost) {
-                navigate(`/post/${dbPost.$id}`);
-            }
-        } else {
-            const file = await appwriteService.uploadFile(data.image[0]);
-
-            if (file) {
-                const fileId = file.$id;
-                data.featuredImage = fileId;
-                const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
+                const dbPost = await appwriteService.updatePost(post.$id, {
+                    ...data,
+                    featuredImage: file ? file.$id : undefined,
+                });
 
                 if (dbPost) {
                     navigate(`/post/${dbPost.$id}`);
                 }
+            } else {
+                if (!data.image[0]) {
+                    console.error("No image provided");
+                    setError("No image provided");
+                    return;
+                }
+                const file = await appwriteService.uploadFile(data.image[0]);
+                console.log("file", file);
+                console.log("userData", userData);
+
+                if (file && userData) {
+                    const fileId = file.$id;
+                    data.featuredImage = fileId;
+                    const dbPost = await appwriteService.createPost({ ...data, userId: userData.$id });
+
+                    if (dbPost) {
+                        navigate(`/post/${dbPost.$id}`);
+                    }
+                }
             }
+        } catch (error) {
+            console.log("Error on PostForm on submit function", error);
+            setError(error.message);
+        } finally {
+            setLoader(false);
         }
     };
 
@@ -60,7 +380,7 @@ export default function PostForm({ post }) {
         return "";
     }, []);
 
-    React.useEffect(() => {
+    useEffect(() => {
         const subscription = watch((value, { name }) => {
             if (name === "title") {
                 setValue("slug", slugTransform(value.title), { shouldValidate: true });
@@ -69,6 +389,10 @@ export default function PostForm({ post }) {
 
         return () => subscription.unsubscribe();
     }, [watch, slugTransform, setValue]);
+
+    if (!userData) {
+        return <div>Loading user data...</div>;
+    }
 
     return (
         <form onSubmit={handleSubmit(submit)} className="flex flex-wrap">
@@ -114,8 +438,15 @@ export default function PostForm({ post }) {
                     {...register("status", { required: true })}
                 />
                 <Button type="submit" bgColor={post ? "bg-green-500" : undefined} className="w-full">
-                    {post ? "Update" : "Submit"}
+                    {loader ? (
+                        <ImSpinner2 className="animate-spin mx-auto" />
+                    ) : post ? (
+                      "Update"
+                    ) : (
+                      "Submit"
+                    )}
                 </Button>
+                {error && <p className="text-red-400">{error}</p>}
             </div>
         </form>
     );
